+++ components/NewProjectButton.tsx
@@ -1,7 +1,7 @@
 import { useRouter } from "next/router";
 import { useMutation } from "react-query";
 import { createProject } from "../services/projectApi";
-import styles from "../pages/processes/Projects.module.scss";
+import styles from "./NewProjectButton.module.scss";
 import { Icon } from "@equinor/eds-core-react";
 import { add } from "@equinor/eds-icons";
 import React from "react";
@@ -16,12 +16,12 @@ export function NewProjectButton(): JSX.Element {
   );
 
   return (
-    <div
+    <button
       className={styles.createCard}
       onClick={() => newProjectMutation.mutate()}
     >
       <Icon data={add} title="add" />
       <p>Create new</p>
-    </div>
+    </button>
   );
 }