+++ interfaces/processLabel.tsx
@@ -1,0 +1,4 @@
+export interface processLabel {
+  id: number;
+  text: string;
+}